Combine all socket clients into RpcSocketClient
