+++ ironfish/src/mining/soloMiner.ts
@@ -237,9 +237,7 @@ export class MiningSoloMiner {
 
     if (!connected) {
       if (!this.connectWarned) {
-        this.logger.warn(
-          `Failed to connect to node on ${String(this.rpc.connection.mode)}, retrying...`,
-        )
+        this.logger.warn(`Failed to connect to node on ${this.rpc.describe()}, retrying...`)
         this.connectWarned = true
       }
